cli, cli/graph: move graph.PlanLines into cli.PlanGraph & cli.VaultGraph
Cleanup the vault & plan display by moving the display logic out of the
inspect subcommand and into the cli package. Split the graph tests into
generic & plan/vault specific.
